requirements.txt
fastapi==0.109.1
uvicorn==0.27.0
python-dotenv==1.0.1
groq>=0.4.1
tensorflow>=2.16.1          # ← upgraded to support Python 3.12
numpy==1.26.4
Pillow==10.2.0
scikit-learn==1.4.0
matplotlib==3.8.2
opencv-python==4.9.0.80
h5py==3.10.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
kaggle==1.5.16
tqdm==4.66.1
scipy==1.12.0

antspyx==0.3.8          # MNI registration
nibabel==5.2.0          # NIfTI file handling
nilearn==0.10.2         # Neuroimaging utilities
xmltodict==0.13.0       # Parse Julich XML